+++ src/data-source.ts
@@ -75,9 +75,12 @@ export class DataSource extends DataSourceWithBackend<RedisQuery, RedisDataSourc
      */
     return {
       ...query,
-      key: query.key ? templateSrv.replace(query.key, scopedVars) : '',
+      keyName: query.keyName ? templateSrv.replace(query.keyName, scopedVars) : '',
       query: query.query ? templateSrv.replace(query.query, scopedVars) : '',
+      field: query.field ? templateSrv.replace(query.field, scopedVars) : '',
       filter: query.filter ? templateSrv.replace(query.filter, scopedVars) : '',
+      legend: query.legend ? templateSrv.replace(query.legend, scopedVars) : '',
+      value: query.value ? templateSrv.replace(query.value, scopedVars) : '',
     };
   }