+++ src/redis/command.ts
@@ -103,7 +103,7 @@ export const CommandParameters = {
   aggregation: ['ts.range', 'ts.mrange'],
   field: ['hget','hmget'],
   filter: ['ts.mrange', 'ts.queryindex'],
-  key: [
+  keyName: [
     'get',
     'hget',
     'hgetall',
@@ -125,6 +125,7 @@ export const CommandParameters = {
   legend: ['ts.range'],
   legendLabel: ['ts.mrange'],
   section: ['info'],
+  value: ['ts.range'],
   valueLabel: ['ts.mrange'],
   fill: ['ts.range', 'ts.mrange'],
   size: ['slowlogGet']