+++ src/tests.ts
@@ -1,0 +1,7 @@
+import Enzyme from 'enzyme';
+import Adapter from 'enzyme-adapter-react-16';
+
+/**
+ * Configure for React 16
+ */
+Enzyme.configure({ adapter: new Adapter() });